Streaming Data Fixes (#1581)
Issue before fix:
* Two issues here that I’m looking to correct. I would like to tread
carefully, however, since this library is used by all teams and the
downstream effects of an error here could be significant.

First: I was noticing that in Respond, the `cancelDestination` call was
being invoked after every successfully completed stream response
against `/respond/incidents`, which is now configured to have a
`cancelDestination`. The problem seems to be that when a socket request
receives a fully completed response, the the `_stopFromSocket` function
is called but the `isStreaming` property is still true, despite the
fact that the request/response transaction is completed. And since
`isStreaming` is true, the logic of the `_stopFromSocket` function is
to invoke `cancelDestination`.

The fix here is simply to set the `isStreaming` property to `false`
whenever the `completed()` method is called. This way,
`_stopFromSocket` processes the `unsubscribe`, but does not attempt a
`cancelDestination` request.

Second: The `stream-cache` has a `_cleanupDuplicateStream` function
that initializes an `allStreams` array by spreading out the anonymous
streams (i.e., streams without a specified `route`). But then in the
following line, it iterates across the keys of `_streams` and pushes
the resulting streams into `allStreams`, including the already handled
anonymous streams so that the anonymous streams get pushed into
`allStreams` array twice, and are duplicated in the that array. Here we
can just remove the initial spread of the anonymous streams into
`allStreams` to avoid that duplication.

Issue: None (discovered when testing canceling stream requests as part
of ASOC-38516)