Respond: Hide Alert Details Nav, show tooltip editable-field (#1590)
Issue before fix:
* The Respond navigation bar appeared on the alert details page, but
does not appear on the incident details page. Need to hide the nav bar
on alert details to make it consistent with the other page.
* The editable-field component can sometimes have long text that is
truncated (and shows ellipsis). We would like to ensure that a standard
`title` appears on this component’s field value element so that the
user can see the full text even when it is truncated (by hovering the
mouse over the field)

What changed
* Added `.respond-engine.alert` selector to css that hides the nav bar,
just like is done on incident details
* Added `title` attribute to editable-field to show tooltip of full text

Issue: ASOC-39869, ASOC-39692