Respond: Display error when microservice is down (#1668)
* Respond: Display error when microservice down

Issue before fix:
* If the respond-server microservice is down or in safe-mode, it
will fail to respond to requests causing an error in the websocket
communication. In the UI, however, there is no error shown, but only
an infinitely spinning load-spinner waiting for a response that will
never come.

What changed:
* In some sense, what we could use is a general health check against the
microservice when the user enters the Respond engine. Alternatively, we could try
to interpret any websocket (or one specific web socket) call error
to represent "service-down."
* In this case, I chose to use the `respond/socket/info` endpoint as a health check.
It is (in theory) a very fast response from the service. By making a "ping" call
against this in the application route, we can ensure the service is up, or if not,
display a Respond server Offline message.
* I wanted to use an error substate for this, but had errors when running in the
`sa` application, even though it appeared fine in the standalone engine dummy app.

Issue: ASOC-35306

* Set respondServerOffline back to false if health check successful

Otherwise, once it's set to true, it will never be unset.