Configure: Respond Incident Rules (#1914)
Issue before fix:
* Originally the Incident Rules (formerly known as Aggregation Rules)
were built into the Respond Engine. But we want them to be hosted under
the new Configure engine instead.
* The name Aggregation Rules was frowned upon, so we're renaming to
Incident Rules.
* Because the Configure engine may host configuration-related components
and routes for other parts of the product, we want to make sure there is
transparency in the routes, components, reducers to reflect those
application parts that are Respond related vs a another part of the
application (e.g., Live Content)

What changed:
* Created a Respond namespace (via folder structure) in Components,
Routes, CSS, Reducers
* Moved Aggregation Rules from Respond engine to the Configure engine's
Respond area.
* Renamed all Aggregation Rules to Incident Rules, except for i18n
references
* Reorganized Reducers/Selectors so that the Incident Rules selectors
are colocated with the reducer, as we see in other parts of the
application. I like this organizational model better than what we have
currently in Respond

Notes:
* A future PR will remove the Aggregation Rule components from Respond
Engine. Currently they still exist.
* A future PR will rename the i18n references
* A future PR will extract shared resources that are used by Respond and
Configure to an addon

Issue: ASOC-45581