Respond: Isolate storyline state and revert to mutable (#2068)
Issue before fix
* We've been seeing regressions caused by seamless immutable's
introduction as it relates to the incident storyline. Because of the
complexity of the underlying code and the use of ember objects in the
selectors, developing a solution that works with seamless-immutable will
take more work than we have availability for at this time. Given that
we want to ensure the feature works properly (as it does in 11.0) the
short-term solution is to do the following:
* First, isolate the storyline state from the incident state, which will
allow us to keep that state completely mutable for the time being (and
thus allow it to work as it originally did). This will involve creating
a new reducer for storyline.
* Rework some of the selectors to bring it back a bit closer to the
original 11.0 version.
* Properly handle the use case for adding a related indicator to the
storyline, now that the storyline reducer functions have been isolated
from the incident reducer functions that include adding a related
indicator to the incident

What changed
* Created a new reducer for the storyline and moved all of the related
code from the incident reducer into this new reducer
* Updated the selectors so that we properly handle the StoryPoint
ember-object
* Updated all references to the storyline state in components that
originally sought for the information in the incident branch of state,
but now needs to pull it from the new storyline branch of state
* Created a promise wrapper around the add-related-indicator api call
promise so that we can pull the related indicator search results from
state (incident-state) and make it available as part of the action
payload to the new storyline-state
* Update tests

Issue: ASOC-46673