Test Helpers - Patch Socket - Throw Socket for method/modelName (#2478)
* Test Helpers - Patch Socket - Throw Socket for method/modelName

Issue before fix
* The `throwSocket` helper has allowed us to simulate a websocket exception on the *next* websocket call. But we do have some components (e.g., `rsa-alerts` and `rsa-incidents` in `Respond` where the component can initiate an action that produces a series of web socket calls, only one of which we want to throw the exception on, and it is not the `next` one.

What changed
* Modified the `patch-socket` `throwSocket` function to allow optional arguments: `methodToThrow` and `modelNameToThrow`. If provided, the exception thrown will only be thrown when `createStream` is invoked with that specified `method` and `modelName`; any other websocket calls with other methods and modelNames will be processed normally without exception.

Issue: None

* Simplify code

* Throw error if createStream is not the original

There was some concern that if the developer calls throwSocket() with the method/modelName parameters, but the socket call is never made, the patch is never reverted. Modified the throwSocket func so that it always checks that the Socket.createStream func is the original, otherwise throws an error. Added a revertPatch() function here so that consumers can revert manually after the test is completed.