Respond: Create Incident Improvements (#2671)
* Respond: Create Incident from Alerts Improvements

Issue before fix:
* A user could select one or more alerts and then create an incident
from them. But the only property on the incident the user could set was
the incident name. We want to be able to set the priority, assignee, and
the category (or categories) all at the time of incident creation.
* Also, we found a defect on the incident details page. Changing the
incident name threw an error. Decided to fix in this PR.

What changed:
* Updated the create-incident component to include assignee, categories,
and priority.
* Updated saga to account for these new details
* Updated english labels
* Added tests
* Fixed defect mentioned above with editing incident name on incident
details page

Issue: ASOC-43960

* Fix: Rename Incident on Incident Details Page

Issue before fix:
* Error thrown when the user tried to rename the incident on the incident details page.

What changed:
* Passed along the required updateItem closure action

Issue: None