Respond: Hide/Show Escalate Button (#2705)
* Respond: Hide/Show Escalate Button

Issue before fix:
* To date, we've shown the Escalate button on the Incidents/Incident
Details pages (in the Inspector) by default. But we only want to show
the button if Archer is configured.
* I ran into an issue with the patch-socket test helper. To date, the
patch-socket helper waits until the stream is stopped before proceeding.
This works fine for all promiseRequests because those are stopped using
the stream.stop function. But streamRequests never get stopped, and
instead they are only ever completed via stream.completed. This was
causing the patch-socket helper to hang and timeout because the expected
stop function was never called

What changed
* Added redux plumbing and API work for fetching incidents settings info
 from respond-server. This tells us whether or not archer is configured
* Updated the incident inspector header and filters components so that
they do not show the Escalate button nor the escalate filters if archer
is not configured
* Added api, selector, and component tests for escalate hide/show
* Updated the patch-socket helper to ensure that the waiter finishes if
stream.completed is called (rather than just stream.stop)

Issue: ASOC-51161, ASOC-51163

* Fix patch-socket update

I wrote this in two phases, and the second phase was a refactor. But the refactor introduced the problem that we patch both funcs even when only one will get called (and restored). Here I'm modifying to always restore the original functions for both, regardless of which func is called.