Refactor: RSA Explorer (#2714)
Issue before fix:
* This is phase one of a multiphased effort to take the already abstract
`rsa-explorer` component used by all three of the Respond list pages
(namely, Incidents, Alerts, and Tasks), and make it even more reusable
as a component in the component-lib. We have another team looking to use
this component in an Admin-related page.

What changed:
* First, removed the action broker, which was responsible for invoking
the correct action-creator. Instead, the consumer will pass the creators
object, which should have  all the requisite explorer-interface action creators in
it, as part of the declaration.
* Next, renamed `namespace` to `reduxSpace`. Namespace was originally
used both to find the correct action-creator method via the broker
(which is no longer the case now that the broker is gone), as well as to
pull the correct stateToComputed properties from redux. This second use
is still in place.
* Removed the `bootstrap` concept from the explorer. Originally, the
consuming component itself was responsible for dispatching all the fetch
actions by implementing the bootstrap action. But I now prefer having
the route dispatch all the fetch calls, and I've created an
`initializeIncidents`, `initializeAlerts`, and an `initializeTasks`
creator that bundles all the fetching calls together, and also optimizes
the fetching now so that we don't refetch information if we already have
it.
* Updated `rsa-explorer` tests and converted the `rsa-incidents`,
`rsa-alerts`, and `rsa-remediation-tasks` tests to vNext.

Notes:
* This is likley phase one of three. The next phase will look at both
i18n and css adjustments to decouple them from anything 'respond'
specific. Then the final phase will be to move the `rsa-explorer` and
tests to the `component-lib` addon so that it can be used by non-respond
teams.

Issue: None