Refactor: RSA Explorer (Part 2) (#2715)
* Respond: RSA-Explorer Part 2

Issue before the fix:
* As phase two (out of likely 3) for the refactoring of rsa-explorer in
preparation for component-izing it, we need to ensure that (a) the class
names and CSS do not reference anything respond related, and (b) all
i18n labels are found outside of the respond label namespace in the
trans-data.js file.

What changed
* Simple reworking of class names and CSS to make the rsa-explorer stuff
independent of respond
* Added a new rsaExplorer section of the i18n translation file and moved
all explorer-related labels under it. Updated all templates to use the
new i18n key(s).
* Updated tests as needed

Notes
* Next phase will be to migrate the rsa-explorer component and
dependencies into the component-lib. It is likely the final phase.

Issues: ASOC-54285

* Fix css in sa app, update docker compose