Component: RSA Explorer  (#2731)
* RSA Explorer Component to component-lib

Issue before fix
* This is the third a final phase of the `component-lib`
componentization of the RSA Explorer (`rsa-explorer`) component that has
been used in Respond as a macro-component.
* The primary objective is to allow other teams to use this component.
* The previous PRs (https://github.rsa.lab.emc.com/asoc/sa-ui/pull/2715
and https://github.rsa.lab.emc.com/asoc/sa-ui/pull/2714) have focused on
refactoring the `rsa-explorer` component within the `respond` engine to
prepare it for migration to the `component-lib`.

What changed
* A lot
* There are two primary sets of changes in this commit: (1) setup of the
`rsa-explorer` component in the `component-lib` along with associated
utilities, mixins, and tests. And (2) updates to `respond` engine, which
includes the removal of the old component, related tests, CSS, and
related utilities, as well as updates to components so that the related
utilities and mixins are referenced from the new `component-lib` space.
The details follow:
* The `rsa-explorer` component files were moved into the
`component-lib`, along with all CSS files, and tests. The necessary app
exports were added, and all components now import the templates and
provide the template as a layout property on the component.
* The explorer initial redux state and reducer functions were added as
utilities
* Since the `rsa-explorer` component is tightly integrated with `redux`,
a reducer and middleware was added to the `component-lib` to accomodate
the `rsa-explorer` tests
* Two mixins (i.e., the `confirmable` and `notifications` mixins) used
both in `respond` and `configure` engines and also used in the
`rsa-explorer` component were migrated into the `component-lib/mixins`
space.
* Test helpers that were required by the `rsa-explorer` tests were moved
to the `test-helpers` addon (or if they were specific to `rsa-explorer`,
added directly to the `component-lib` test helpers directory).
* In `respond`, all references to explorer-related resources were
updated to use those now found in the `component-lib`. There was some
improvement as well in the use of the `notifications` mixin to reduce
the amount of code needed.
* In `configure`, the duplicate references (e.g., `confirmable` mixin)
were removed and now reference the `component-lib` resource instead
* All `rsa-explorer` components, CSS, tests were removed from the
`respond` engine.

Notes:
* Currently there is no documentation around the `rsa-explorer`
component in the `style-guide`. This can come later, if desired.

Issue:
* ASOC-54285

* Fix ESLint error