Respond: Toggle Entity Node Visibility (#2791)
Issue before fix:
* Some nodal visualizations in Respond can be very complex, which can
make them difficult to read. We want to improvement the UX by allowing
them to hide/show nodes based on specific entity types so that the user
can focus on only those entities that are of interest and reduce the
visual complexity of the nodal graph.
* We had some old (dead) code from the original nodal graph that could
be safely removed

What changed
* Removed old/dead code and tests from the original (declarative) nodal
graph. This code had some performance problems at scale, which is why it was
replaced with the "rsa-fast-force" component (imperative d3).
* Updated the nodal graph's entities legend to add checkboxes next to
the label
* Updated reducer, action types, creators for setting the default
visibility of entity types (all visible) and toggling them off and back
on.
* Updated the nodal graph code to add additional classes to links and
nodes so that we can hide (using CSS) the nodes and the links from or
to nodes that are not supposed to be visible.
* Updated CSS so that it does not display hidden nodes/links, and we
increased the size of the entities legend font to make this a bit more
visible.
* Added tests for components, selectors, reducers.

Issue: ASOC-44399