Configure - Log Parsers: Add Parser (#2937)
* Configure - Log Parsers: Add Parser

Issue before fix:
* There was no way to add a new log parser
* The Add Parser controls appear in a modal when the add-new button is
clicked in the Parsers toolbar. The form shows a set of dropdowns and
fields. Selecting a pre-existing (non-New) parser will autopopulated all
the fields except for the Clone From dropdown. These fields are
disabled. If the New option is used, the user can edit the fields and
create a brand new parser.

What changed:
* Added a new button to the Parser toolbar: Add New (Parser)
* Added new component `log-parsers/add-log-parser` which represents the
form that appears in a modal for adding a new log parser
* Added API/Creators/Reducers/Selectors for fetching and selecting
device-classes and device-types, which are used in the add-log-parser
form
* Added data for mock-server
* Added CSS
* Added tests

Notes:
* No tests just yet for the toolbar in terms of testing the opening of
the modal/closing of the modal. Those will follow in the next day.
Otherwise, there are full tests on the Add Parser form, where the
general complexity lies.

Issue: ASOC-41949

* Simplify test

* Skipping test for now

This test runs fine locally but not in jenkins. Skipping for now to move forward.