Configure - Log Parsers - Sample Log Highlighting (#2989)
* Configure - Log Parsrs - Sample Log Highlighting

Issue before fix:
* We had hardcoded text in the sample log section of the log parsers
page. There was no highlighting of the text based on the parser rules.

What changed:
* Added API/Reducers/Selectors for highlighting
* Added component file for the `log-message` component
* Added CSS
* Added tests for component, selectors, reducers

Notes:
* We need to detect paste events to properly strip html from the
clipboard text. Added an ember customEvents hash to the app to activate
this event so that we can handle in components.
* Because Glimmer tries to be smart about rendering, and because
contenteditable elements add additional html elements while the user
types, I needed some extra handling around rendering the `pre` tag so
that we replace the entire `pre` element before state changes cause
glimmer to re-evaluate the dom. Otherwise we get DOM node removal errors
when Glimmer finds DOM elements it wasn't expecting. Now, when the highlighting operation is in
progress, we swap in a new `pre` tag with only text (no html) to ensure
that when the highlighting response returns, we will re-render the
entire `pre` element without any fancy glimmer optimizations.
* We still need to hook up the action creator for highlighting to
dispatch whenever changes are made to the tokens or other parts of the
rule.

Issue: Unknown

* Fix selector test